requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiomysql==0.2.0
SQLAlchemy[asyncio]==2.0.30
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
